+++ api/resources_portal/test/views/importers/test_geo.py
@@ -1,9 +1,17 @@
+from unittest.mock import patch
+
 from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APITestCase
 
 from resources_portal.models import Material
-from resources_portal.test.factories import GrantFactory, OrganizationFactory, UserFactory
+from resources_portal.test.factories import (
+    GrantFactory,
+    MaterialFactory,
+    OrganizationFactory,
+    UserFactory,
+)
+from resources_portal.test.utils import get_mock_dataset_data
 
 
 class ImportGEOTestCase(APITestCase):
@@ -49,7 +57,7 @@ class ImportGEOTestCase(APITestCase):
             material.additional_metadata["accession_code"], self.test_accession_with_pubmed_id
         )
         self.assertEqual(
-            material.additional_metadata["number_samples"],
+            material.additional_metadata["number_of_samples"],
             self.test_accession_with_pubmed_id_num_samples,
         )
 
@@ -76,9 +84,35 @@ class ImportGEOTestCase(APITestCase):
             material.additional_metadata["accession_code"], self.test_accession_without_pubmed_id
         )
         self.assertEqual(
-            material.additional_metadata["number_samples"],
+            material.additional_metadata["number_of_samples"],
             self.test_accession_without_pubmed_id_num_samples,
         )
+
+    @patch("resources_portal.importers.geo.gather_all_metadata", side_effect=get_mock_dataset_data)
+    def test_import_previously_imported_geo_fails(self, mock_dataset):
+        self.client.force_authenticate(user=self.user)
+        accession_code = "12345"
+        response = self.client.post(
+            self.url, {"import_source": "GEO", "accession_code": accession_code}
+        )
+        import pdb
+
+        pdb.set_trace()
+        MaterialFactory(
+            imported=True,
+            import_source="GEO",
+            additional_metadata={
+                "accession_code": response.json()["additional_metadata"]["accession_code"]
+            },
+        )
+
+        # Try to import the same material again
+        response = self.client.post(
+            self.url, {"import_source": "GEO", "accession_code": accession_code}
+        )
+
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(response.json()["error_code"], "ALREADY_IMPORTED")
 
     def test_import_from_unauthenticated_fails(self):
         url = reverse("materials-import")
