+++ api/resources_portal/test/views/importers/test_protocol_io.py
@@ -1,9 +1,17 @@
+from unittest.mock import patch
+
 from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APITestCase
 
 from resources_portal.models import GrantOrganizationAssociation, Material
-from resources_portal.test.factories import GrantFactory, OrganizationFactory, UserFactory
+from resources_portal.test.factories import (
+    GrantFactory,
+    MaterialFactory,
+    OrganizationFactory,
+    UserFactory,
+)
+from resources_portal.test.utils import get_mock_protocol_data
 
 
 class ImportProtocolTestCase(APITestCase):
@@ -79,6 +87,32 @@ class ImportProtocolTestCase(APITestCase):
         )
 
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+    @patch(
+        "resources_portal.importers.protocols_io.gather_all_metadata",
+        side_effect=get_mock_protocol_data,
+    )
+    def test_import_previously_imported_protocol_fails(self, mock_protocol):
+        self.client.force_authenticate(user=self.user)
+        protocol_doi = "12345"
+        response = self.client.post(
+            self.url, {"import_source": "PROTOCOLS_IO", "protocol_doi": protocol_doi}
+        )
+        MaterialFactory(
+            imported=True,
+            import_source="PROTOCOLS_IO",
+            additional_metadata={
+                "protocol_doi": response.json()["additional_metadata"]["protocol_doi"]
+            },
+        )
+
+        # Try to import the same material again
+        response = self.client.post(
+            self.url, {"import_source": "PROTOCOLS_IO", "protocol_doi": protocol_doi}
+        )
+
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(response.json()["error_code"], "ALREADY_IMPORTED")
 
     def test_import_from_unauthenticated_fails(self):
         response = self.client.post(
